requirements.txt
# Django y framework
Django==4.2.0
djangorestframework==3.14.0

# Machine Learning
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2

# Gráficas
matplotlib==3.7.2
seaborn==0.12.2

# Producción
gunicorn==21.2.0
whitenoise==6.5.0
python-dotenv==1.0.0
dj-database-url==2.0.0
psycopg2-binary==2.9.7  # Para PostgreSQL

# Seguridad
cryptography==41.0.5